move youtube track to own file
