+++ src/player/track.ts
@@ -1,75 +1,39 @@
-import { stringify } from "querystring";
 import { Readable } from "stream";
-import ytdl, {
-  chooseFormatOptions as ChooseFormatOptions,
-  videoFormat as VideoFormat,
-  videoInfo as VideoInfo,
-} from "ytdl-core";
-import { msToDuration } from "../util/string";
+import { videoFormat as VideoFormat } from "ytdl-core";
 
 export interface Track {
+  /**
+   * The string representation for the track
+   */
   toString(): string;
+
+  /**
+   * A playable source.
+   */
   getSource(): Promise<string | Readable>;
-  supports(query: string): Promise<boolean>;
+
+  /**
+   * @todo Is this used anywhere?
+   */
   getFormat(): VideoFormat;
+
+  /**
+   *
+   */
   duration: string;
+
+  /**
+   * @todo redundant?
+   */
   durationMs: number;
-  url: string;
-  thumbnail: string;
-}
 
-export class YoutubeTrack implements Track {
-  private format: VideoFormat;
+  /**
+   * The web link (if any)
+   */
+  url?: string;
 
-  private static formatOptions: ChooseFormatOptions = {
-    quality: "highestaudio",
-  };
-
-  constructor(private info: VideoInfo) {
-    this.format = ytdl.chooseFormat(
-      this.info.formats,
-      YoutubeTrack.formatOptions,
-    );
-  }
-
-  /** @todo catch errors */
-  static async fromUrl(url: string) {
-    const info = await ytdl.getInfo(url);
-    return new YoutubeTrack(info);
-  }
-
-  toString() {
-    return this.info.videoDetails.title;
-  }
-
-  async getSource() {
-    return ytdl(this.url, YoutubeTrack.formatOptions);
-  }
-
-  getFormat() {
-    return this.format;
-  }
-
-  get durationMs() {
-    return this.format?.approxDurationMs
-      ? parseInt(this.format.approxDurationMs)
-      : 0;
-  }
-
-  get duration() {
-    return msToDuration(this.durationMs);
-  }
-
-  /** @todo */
-  async supports(query: string) {
-    return true;
-  }
-
-  get url() {
-    return this.info.videoDetails.video_url;
-  }
-
-  get thumbnail() {
-    return this.info.videoDetails.thumbnails?.[0].url;
-  }
+  /**
+   * The track thumbnail
+   */
+  thumbnail?: string;
 }
